+++ eyes.sdk.core/src/main/java/com/applitools/eyes/fluent/ICheckSettingsInternal.java
@@ -39,4 +39,6 @@ public interface ICheckSettingsInternal extends com.applitools.ICheckSettingsInt
     GetAccessibilityRegion[] getAccessibilityRegions();
 
     BaseOcrRegion getOcrRegion();
+
+    String getVariationGroupId();
 }