+++ eyes.selenium.java/src/main/java/com/applitools/eyes/selenium/rendering/VisualGridEyes.java
@@ -210,6 +210,7 @@ public class VisualGridEyes implements ISeleniumEyes {
 
         List<VisualGridRunningTest> newTests = new ArrayList<>();
         ServerConnector serverConnector = runner.getServerConnector();
+        String agentRunId = String.format("%s_%s", getConfiguration().getTestName(), UUID.randomUUID().toString());
         for (RenderBrowserInfo browserInfo : browserInfoList) {
             if (browserInfo.getEmulationInfo() != null) {
                 Map<String, DeviceSize> deviceSizes = serverConnector.getDevicesSizes(ServerConnector.EMULATED_DEVICES_PATH);
@@ -219,7 +220,7 @@ public class VisualGridEyes implements ISeleniumEyes {
                 Map<String, DeviceSize> deviceSizes = serverConnector.getDevicesSizes(ServerConnector.IOS_DEVICES_PATH);
                 browserInfo.setIosDeviceSize(deviceSizes.get(browserInfo.getIosDeviceInfo().getDeviceName()));
             }
-            VisualGridRunningTest test = new VisualGridRunningTest(logger, eyesId, getConfiguration(), browserInfo, this.properties, serverConnector);
+            VisualGridRunningTest test = new VisualGridRunningTest(logger, eyesId, getConfiguration(), browserInfo, this.properties, serverConnector, agentRunId);
             this.testList.put(test.getTestId(), test);
             newTests.add(test);
         }